refactor: RelativeStatsViewModel을 상대전적, 상대전적 디테일 프래그먼트에서 함께 사용하도록 변경
