+++ android/app/src/main/java/com/foss/foss/feature/home/HomeActivity.kt
@@ -22,7 +22,6 @@ class HomeActivity : DiActivity() {
 
     private val recentMatchesViewModel: RecentMatchesViewModel by diViewModels()
     private val relativeStatsViewModel: RelativeStatsViewModel by diViewModels()
-    private val relativeDetailStatsViewModel: RelativeDetailStatsViewModel by diViewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -31,7 +30,6 @@ class HomeActivity : DiActivity() {
         setupHomeView()
         setupRecentMatchesObserver()
         setupRelativeStatsObserver()
-        setUpRelativeDetailStatsObserver()
         setSearchingMatchesButtonClickListener()
     }
 
@@ -78,15 +76,6 @@ class HomeActivity : DiActivity() {
     private fun setupRelativeStatsObserver() {
         repeatOnStarted {
             relativeStatsViewModel.relativeStats.collect {
-            }
-        }
-    }
-
-    private fun setUpRelativeDetailStatsObserver() {
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                relativeDetailStatsViewModel.matches.collect {
-                }
             }
         }
     }
